♻️ Refactoring popups
